Remove controller behavior modules
This pattern was established in order to allow local applications to
override the upstream gem classes.  In ruby 2 we can do this using
Module#prepend instead.

I believe that merging these files will make the development effort
simpler.

Ref #799
