+++ .rubocop.yml
@@ -66,7 +66,7 @@ Style/MethodMissing:
 
 Style/NumericPredicate:
   Exclude:
-    - 'app/controllers/concerns/hyrax/file_sets_controller_behavior.rb'
+    - 'app/controllers/hyrax/file_sets_controller.rb'
 
 Style/SymbolArray:
   Enabled: false
@@ -91,7 +91,7 @@ Style/SingleLineBlockParams:
 
 Style/ZeroLengthPredicate:
   Exclude:
-    - 'app/controllers/concerns/hyrax/file_sets_controller_behavior.rb'
+    - 'app/controllers/hyrax/file_sets_controller.rb'
 
 Rails:
   Enabled: true
