+++ app/controllers/hyrax/uploads_controller.rb
@@ -1,5 +1,16 @@
 module Hyrax
   class UploadsController < ApplicationController
-    include Hyrax::UploadsControllerBehavior
+    load_and_authorize_resource class: Hyrax::UploadedFile
+
+    def create
+      @upload.attributes = { file: params[:files].first,
+                             user: current_user }
+      @upload.save!
+    end
+
+    def destroy
+      @upload.destroy
+      head :no_content
+    end
   end
 end
